# Handle incompatible DeepSeek GGUFs