@ikawrakow 2025-05-09
> WARNING:gguf.vocab:Adding merges requested but no merges found, output may be non-functional.

Yes, the `convert_hf_to_gguf.py` script currently on master does not handle merges well. There is a fix in PR #377, but I haven't merged because for some reason it misses the `rope_scaling` tensor, and we have not understood why.