@Panchovix 2025-05-09
I'm testing now! With DeepSeekV3 0324 Q2_K_XL latest quant, on 128GB VRAM (5090+4090x2+A6000) and 192GB RAM (6000Mhz 7800X3D). But first I just noticed this

```
llm_load_tensors:        CPU buffer size = 133756.00 MiB
llm_load_tensors:  CUDA_Host buffer size =   497.11 MiB
llm_load_tensors:      CUDA0 buffer size = 22412.07 MiB
llm_load_tensors:      CUDA1 buffer size = 17714.47 MiB
llm_load_tensors:      CUDA2 buffer size = 21610.08 MiB
llm_load_tensors:      CUDA3 buffer size = 42786.36 MiB
```

Is there a way to load on GPU first and then CPU? This explains why on ikllamacpp I get 5-20 t/s on PP vs 60-100 t/s on llamacpp (on the latter it looks like this)

```
load_tensors: offloading output layer to GPU
load_tensors: offloaded 62/62 layers to GPU
load_tensors:        CUDA0 model buffer size = 22412.07 MiB
load_tensors:        CUDA1 model buffer size = 17714.47 MiB
load_tensors:        CUDA2 model buffer size = 21610.08 MiB
load_tensors:        CUDA3 model buffer size = 42786.36 MiB
load_tensors:          CPU model buffer size = 134253.11 MiB
```

Okay now regarding the model itself, I have loaded it with (no fa since I think fa is merged on main but not on the PR), had to change the allocation a bit to make it work.

```
./llama-server -m '/GGUFs/DeepSeek-V3-0324-UD-Q2_K_XL-merged.gguf' -c 16384 --no-mmap --no-warmup -v -ngl 999 -ot "blk.(0|1|2|3|4|5|6).ffn.=CUDA0" -ot "blk.(7|8|9|10).ffn.=CUDA1" -ot "blk.(11|12|13|14).ffn.=CUDA2"  -ot "blk.(15|16|17|18|19|20|21|22|23|24).ffn.=CUDA3"  -ot "ffn.*=CPU" -fmoe -mla 1

```

And it loads without issues

```
llm_load_tensors: offloading 61 repeating layers to GPU
llm_load_tensors: offloading non-repeating layers to GPU
llm_load_tensors: offloaded 62/62 layers to GPU
llm_load_tensors:        CPU buffer size = 142690.30 MiB
llm_load_tensors:  CUDA_Host buffer size =   497.11 MiB
llm_load_tensors:      CUDA0 buffer size = 18265.88 MiB
llm_load_tensors:      CUDA1 buffer size = 17471.11 MiB
llm_load_tensors:      CUDA2 buffer size = 17472.86 MiB
llm_load_tensors:      CUDA3 buffer size = 42378.83 MiB
....................................................................................................
llama_new_context_with_model: n_ctx      = 16384
llama_new_context_with_model: n_batch    = 2048
llama_new_context_with_model: n_ubatch   = 512
llama_new_context_with_model: flash_attn = 0
llama_new_context_with_model: mla_attn   = 1
llama_new_context_with_model: attn_max_b = 0
llama_new_context_with_model: fused_moe  = 1
llama_new_context_with_model: ser        = -1, 0
llama_new_context_with_model: freq_base  = 10000.0
llama_new_context_with_model: freq_scale = 0.025
llama_kv_cache_init:      CUDA0 KV buffer size =   510.00 MiB
llama_kv_cache_init:      CUDA1 KV buffer size =   408.00 MiB
llama_kv_cache_init:      CUDA2 KV buffer size =   408.00 MiB
llama_kv_cache_init:      CUDA3 KV buffer size =   748.00 MiB
llama_new_context_with_model: KV self size  = 2074.00 MiB, c^KV (f16): 1098.00 MiB, kv^T (f16):  976.00 MiB
llama_new_context_with_model:  CUDA_Host  output buffer size =     0.99 MiB
llama_new_context_with_model: pipeline parallelism enabled (n_copies=1)
llama_new_context_with_model:      CUDA0 compute buffer size =  4522.00 MiB
llama_new_context_with_model:      CUDA1 compute buffer size =  4481.00 MiB
llama_new_context_with_model:      CUDA2 compute buffer size =  4481.00 MiB
llama_new_context_with_model:      CUDA3 compute buffer size =  4481.00 MiB
llama_new_context_with_model:  CUDA_Host compute buffer size =    78.50 MiB
llama_new_context_with_model: graph nodes  = 3547
llama_new_context_with_model: graph splits = 398
```

Then generating also works without issues! 

Speeds look like this

```
INFO [           print_timings] prompt eval time     =  246500.09 ms /  3218 tokens (   76.60 ms per token,    13.05 tokens per second) | tid="140049526018048" timestamp=1746811228 id_slot=0 id_task=0 t_prompt_processing=246500.088 n_prompt_tokens_processed=3218 t_token=76.60040024860162 n_tokens_second=13.05476207375634
INFO [           print_timings] generation eval time =   63970.82 ms /   428 runs   (  149.46 ms per token,     6.69 tokens per second) | tid="140049526018048" timestamp=1746811228 id_slot=0 id_task=0 t_token_generation=63970.815 n_decoded=428 t_token=149.46452102803738 n_tokens_second=6.690550995793941
INFO [           print_timings]           total time =  310470.90 ms | tid="140049526018048" timestamp=1746811228 id_slot=0 id_task=0 t_prompt_processing=2
```

For reference, as I mentioned above on llamacpp with the same command but having CUDA0 loading first instead of CPU, I get

```
prompt eval time = 51369.66 ms / 3252 tokens ( 15.80 ms per token, 63.31 tokens per second)
```

So I can confirm latest quants with MLA works on ik llamacpp.