@ikawrakow 2025-05-09
@Panchovix Thanks for testing!

Why don't you simply use the same tensor overrides that you use with mainline `llama.cpp`?

If you post your `llama.cpp` command here, perhaps we can give you suggestions how you can improve it for `ik_llama.cpp`.